Add support for new "dont_use_session_level_serverhost_attribute" config option.
This option can be used in combination with event level
"__origServerHost" attribute.

Technically we should really always use "__origServerHost" attribute
internally since the way it's currently implemented, event level
"serverHost" attribute is a regulat event attribute (like any other
event level attribute) so it doesn't provide much value.

But that change would technically be more invasive so I decided to go
with this opt-in approach to begin with.