Add support for setting special serverHost attribute on for events in a batch which don't explicitly define this special attribute to a server level value.
This is needed in case at least one event in the batch defines this
special attribute. In such scenario, we don't send session level
serverHost since it would overwrite event level one. Which means events
which don't explicitly define serverHost will contain no server host.

This change supports for setting serverHost on such events to the value
defined on the session level.

This change requires additional post-processing and includes non-trivial
amount of overhead so it's opt-in and disabled by default.

In most cases, default behavior should be good enough (since usually if
serverHost is defined on a filter on the event, it will usually be
defined on all the events in a batch).