Make sure we still report status for "active" plugins even if we haven't received data for a while.
We perform total events processes count check to ensure we only report
status for plugins which received some data.

In a lot of setups, each logstash if configured with a bunch of idle
plugin instances which don't receive any data at all and we don't want
to report stauts for those instances.