Change "Unexpected error occured while uploading to Scalyr" log message to use INFO and not error log level.
This message is not level and simply just indicates that a client will
retry failed attempt. In case all the retry attempts are exhaused and we
still fail to send data, error message is already logged.

Technically we could also use INFO, but a lot of deployments have log
level set to >= WARNING and with INFO we would lose visibility into
those messages.