Update the code to not set Event.sev attribute in case the severity field value is not valid (not a numeric value between 0 and 6).
In case DataSet API sees invalid severity value, it will not process the
batch and return 400 error to the client.

This change makes sure the field only gets set if the value is valid -
this way we make sure server won't drop the value in case an invalid
value is provided by the logstash operator or similar.