Implement exception type specific HTTP request retrying.
This allows us to reduce unnecessary load on the client and on the server because we won't retry
so often in scenarios where that doesn't make sense - e.g. during deploy windows when we expect a
slightly longer unavailability and when client is throttled by the server.

On the other hand, we still want to initially retry quite fast on other errors (e.g. short network
hiccup or similar).