Bump setup-ruby action version to fix bundler install failure
The GHA workflows using jruby are failing when installing bundler during
the setup-ruby action with the following error:

ERROR:  Error installing bundler: The last version of bundler (~> 2.0)
to support your Ruby & RubyGems was 2.3.26. Try installing it with `gem
install bundler -v 2.3.26`

The version of bundler specified in the lockfile is 2.3.22, and recent
versions of the setup-ruby action will install the bundler version
specified in the lockfile.

DSET-2151